feat: Day 12 done.
This took much longer than I expected, but I'm happing with the timings.

Ended up stealing the `pathfinding` crate to find the regions. I may end
up rewriting the part of it to use the `hashbrown` crate instead of
`std::collections::HashSet`.

And it was silly to need to store the
starts in a `Vec` instead of just using an `IntoIterator`. Especially
because they call `into_iter` on the `Vec` anyway.